+++ sugar/components/server/cdatamatch.py
@@ -29,7 +29,7 @@ import re
 
 import sugar.utils.objects
 import sugar.utils.structs
-from sugar.components.server.cdatastore import CDataContainer
+from sugar.components.server.pdatastore import CDataContainer
 from sugar.components.server.qelement import QueryBlock
